Stop using spinlocks
It turns out that spinlocks are actually unsafe on iOS, because the
system does not guarantee progress in the face of priority inversion
(not even OSSpinLock suffices). Since the system doesn't provide us with
a safe spinlock, we have to do things a different way.

To protect the _callback, introduce an "activity count", which is
basically a retain count for the ivar. This guarantees that anyone
reading the ivar can do so safely, and the ivar will be nilled out once
the count hits zero.

To block in -teardown, switch to a pthread mutex. Not only is this safe,
it also automatically handles priority inversion by increasing the
priority of the thread with the mutex. We also introduce a new flag into
_state that lets us tell ahead of time if we can skip blocking.

Also go through and ditch seq_cst memory ordering. We need a
synchronizes-with edge between init and teardown, but we can use relaxed
everywhere else.