Work around compiler bug with properties of type `Any?`
Swift 5.1.5 and above have a bug with the expression
`Value.self is Any.Protocol`. In Swift 5.1.5 this returns the incorrect
results under debug and a modern deployment target, but the correct
results under optimization or with an older deployment target. In Swift
5.2 this returns the incorrect results in debug with all deployment
targets, in addition to having an incorrect compiler warning.

Fixes #25.